Move build targets into separate files
