+++ lib/build_targets/application.dart
@@ -1,0 +1,174 @@
+// Copyright 2021 Samsung Electronics Co., Ltd. All rights reserved.
+// Copyright 2014 The Flutter Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+// @dart = 2.8
+
+import 'package:file/file.dart';
+import 'package:flutter_tools/src/artifacts.dart';
+import 'package:flutter_tools/src/base/file_system.dart';
+import 'package:flutter_tools/src/build_info.dart';
+import 'package:flutter_tools/src/build_system/build_system.dart';
+import 'package:flutter_tools/src/build_system/depfile.dart';
+import 'package:flutter_tools/src/build_system/targets/assets.dart';
+import 'package:flutter_tools/src/build_system/targets/icon_tree_shaker.dart';
+import 'package:flutter_tools/src/build_system/exceptions.dart';
+import 'package:flutter_tools/src/build_system/source.dart';
+import 'package:flutter_tools/src/build_system/targets/android.dart';
+import 'package:flutter_tools/src/build_system/targets/common.dart';
+
+import '../tizen_builder.dart';
+import 'plugins.dart';
+
+/// Prepares the pre-built Flutter bundle.
+///
+/// Source: [AndroidAssetBundle] in `android.dart`
+abstract class TizenAssetBundle extends Target {
+  const TizenAssetBundle();
+
+  @override
+  String get name => 'tizen_asset_bundle';
+
+  @override
+  List<Source> get inputs => const <Source>[
+        Source.pattern('{BUILD_DIR}/app.dill'),
+        ...IconTreeShaker.inputs,
+      ];
+
+  @override
+  List<Source> get outputs => const <Source>[];
+
+  @override
+  List<String> get depfiles => <String>[
+        'flutter_assets.d',
+      ];
+
+  @override
+  List<Target> get dependencies => const <Target>[
+        KernelSnapshot(),
+      ];
+
+  @override
+  Future<void> build(Environment environment) async {
+    if (environment.defines[kBuildMode] == null) {
+      throw MissingDefineException(kBuildMode, name);
+    }
+    final BuildMode buildMode =
+        getBuildModeForName(environment.defines[kBuildMode]);
+    final Directory outputDirectory = environment.outputDir
+        .childDirectory('flutter_assets')
+          ..createSync(recursive: true);
+
+    // Only copy the prebuilt runtimes and kernel blob in debug mode.
+    if (buildMode == BuildMode.debug) {
+      final String vmSnapshotData = environment.artifacts
+          .getArtifactPath(Artifact.vmSnapshotData, mode: BuildMode.debug);
+      final String isolateSnapshotData = environment.artifacts
+          .getArtifactPath(Artifact.isolateSnapshotData, mode: BuildMode.debug);
+      environment.buildDir
+          .childFile('app.dill')
+          .copySync(outputDirectory.childFile('kernel_blob.bin').path);
+      environment.fileSystem
+          .file(vmSnapshotData)
+          .copySync(outputDirectory.childFile('vm_snapshot_data').path);
+      environment.fileSystem
+          .file(isolateSnapshotData)
+          .copySync(outputDirectory.childFile('isolate_snapshot_data').path);
+    }
+    final Depfile assetDepfile = await copyAssets(
+      environment,
+      outputDirectory,
+      targetPlatform: null, // corresponds to flutter-tester
+    );
+    final DepfileService depfileService = DepfileService(
+      fileSystem: environment.fileSystem,
+      logger: environment.logger,
+    );
+    depfileService.writeToFile(
+      assetDepfile,
+      environment.buildDir.childFile('flutter_assets.d'),
+    );
+  }
+}
+
+/// Generates an AOT snapshot (app.so) of the Dart code.
+///
+/// Source: [AotElfRelease] in `common.dart`
+class TizenAotElf extends AotElfBase {
+  TizenAotElf();
+
+  @override
+  String get name => 'tizen_aot_elf';
+
+  @override
+  List<Source> get inputs => <Source>[
+        const Source.pattern('{BUILD_DIR}/app.dill'),
+        const Source.artifact(Artifact.engineDartBinary),
+        const Source.artifact(Artifact.skyEnginePath),
+        // Any type of gen_snapshot is applicable here because engine artifacts
+        // are assumed to be updated at once, not one by one for each platform
+        // or build mode.
+        const Source.artifact(Artifact.genSnapshot, mode: BuildMode.release),
+      ];
+
+  @override
+  List<Source> get outputs => const <Source>[
+        Source.pattern('{BUILD_DIR}/app.so'),
+      ];
+
+  @override
+  List<Target> get dependencies => const <Target>[
+        KernelSnapshot(),
+      ];
+}
+
+/// Source: [DebugAndroidApplication] in `android.dart`
+class DebugTizenApplication extends TizenAssetBundle {
+  DebugTizenApplication(this.buildInfo);
+
+  final TizenBuildInfo buildInfo;
+
+  @override
+  String get name => 'debug_tizen_application';
+
+  @override
+  List<Source> get inputs => <Source>[
+        ...super.inputs,
+        const Source.artifact(Artifact.vmSnapshotData, mode: BuildMode.debug),
+        const Source.artifact(Artifact.isolateSnapshotData,
+            mode: BuildMode.debug),
+      ];
+
+  @override
+  List<Source> get outputs => <Source>[
+        ...super.outputs,
+        const Source.pattern('{OUTPUT_DIR}/flutter_assets/vm_snapshot_data'),
+        const Source.pattern(
+            '{OUTPUT_DIR}/flutter_assets/isolate_snapshot_data'),
+        const Source.pattern('{OUTPUT_DIR}/flutter_assets/kernel_blob.bin'),
+      ];
+
+  @override
+  List<Target> get dependencies => <Target>[
+        ...super.dependencies,
+        NativePlugins(buildInfo),
+      ];
+}
+
+/// See: [ReleaseAndroidApplication] in `android.dart`
+class ReleaseTizenApplication extends TizenAssetBundle {
+  ReleaseTizenApplication(this.buildInfo);
+
+  final TizenBuildInfo buildInfo;
+
+  @override
+  String get name => 'release_tizen_application';
+
+  @override
+  List<Target> get dependencies => <Target>[
+        ...super.dependencies,
+        TizenAotElf(),
+        NativePlugins(buildInfo),
+      ];
+}
